feat(i18n): add internationalization support with English and Chinese translations
- Implemented i18n functionality using vue-i18n.
- Added English locale translations in src/i18n/locales/en.js.
- Added Chinese locale translations in src/i18n/locales/zh-CN.js.
- Created i18n configuration file in src/i18n/index.js.
- Updated yarn.lock to include new dependencies for esbuild and rollup.
