+++ src/i18n/locales/zh-CN.js
@@ -1,0 +1,77 @@
+export default {
+  message: {
+    hello: '你好',
+    welcome: '欢迎使用ShotMeld'
+  },
+  settings: {
+    title: '设置',
+    subtitle: '自定义您的 ShotMeld 体验',
+    theme: {
+      title: '主题',
+      description: '选择您喜欢的界面主题',
+      light: '浅色模式',
+      dark: '深色模式',
+      system: '跟随系统'
+    },
+    language: {
+      title: '语言',
+      description: '选择您偏好的界面语言',
+      zh: '中文',
+      en: 'English'
+    }
+  },
+  profile: {
+    loading: '正在加载个人信息...',
+    email: '邮箱',
+    registerDate: '注册时间',
+    photos: '照片',
+    noPhotos: '无照片',
+    albums: '相册',
+    noAlbums: '未创建',
+    logout: '退出登录',
+    fetchError: '获取用户信息失败:'
+  },
+  sharePhoto: {
+    loading: '正在加载照片...',
+    error: {
+      title: '照片不存在或未分享',
+      homeButton: '返回首页',
+      invalidId: '照片ID无效',
+      loadFailed: '加载共享照片失败:',
+      notFound: '照片不存在或未分享',
+      imageLoadFailed: '照片加载失败'
+    },
+    defaultTitle: '共享照片',
+    download: '下载',
+    imageLoading: '正在加载图片...',
+    pageTitle: 'Shotmeld 共享照片',
+    timeline: {
+      loading: '正在加载时间轴...',
+      retry: '重试',
+      year: '年',
+      month: '月',
+      photos: '张照片',
+      noTitle: '无标题',
+      uploadModal: {
+        title: '上传照片'
+      },
+      error: {
+        fetchFailed: '获取照片失败:',
+        loadFailed: '加载时间轴失败，请重试',
+        deleteFailed: '删除照片失败:',
+        deleteFailedTitle: '删除失败',
+        deleteFailedMessage: '无法删除照片，请重试'
+      },
+      editPhoto: '编辑照片:',
+      editPhotoComingSoon: '编辑照片功能将在后续版本实现',
+      uploadSuccess: {
+        title: '上传成功',
+        message: '已成功上传 {count} 张照片'
+      },
+      deleteSuccess: {
+        title: '成功',
+        message: '照片已删除'
+      }
+    }
+  }
+}
